+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/Customers/DeleteCustomer.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using AzureFunctions.TestApplication.Application.Customers.DeleteCustomer;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
 using AzureFunctions.TestApplication.Domain.Common.Interfaces;