+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/ListedUnlistedServices/ListedServiceFunc.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using AzureFunctions.TestApplication.Application.Interfaces;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
 using AzureFunctions.TestApplication.Domain.Common.Interfaces;
@@ -45,9 +46,15 @@ namespace AzureFunctions.TestApplication.Api
             try
             {
                 string param = req.Query["param"];
-                await _appService.ListedServiceFunc(param);
-                await _unitOfWork.SaveChangesAsync(cancellationToken);
-                return new CreatedResult(string.Empty, null);
+
+                using (var transaction = new TransactionScope(TransactionScopeOption.Required,
+                    new TransactionOptions() { IsolationLevel = IsolationLevel.ReadCommitted }, TransactionScopeAsyncFlowOption.Enabled))
+                {
+                    await _appService.ListedServiceFunc(param, cancellationToken);
+                    await _unitOfWork.SaveChangesAsync(cancellationToken);
+                    transaction.Complete();
+                    return new CreatedResult(string.Empty, null);
+                }
             }
             catch (NotFoundException exception)
             {