+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/SampleDomainsService/FindSampleDomains.cs
@@ -20,28 +20,26 @@ using Microsoft.OpenApi.Models;
 
 namespace AzureFunctions.TestApplication.Api
 {
-    public class FindSampleDomainById
+    public class FindSampleDomains
     {
         private readonly ISampleDomainsService _appService;
 
-        public FindSampleDomainById(ISampleDomainsService appService)
+        public FindSampleDomains(ISampleDomainsService appService)
         {
             _appService = appService ?? throw new ArgumentNullException(nameof(appService));
         }
 
-        [FunctionName("FindSampleDomainById")]
-        [OpenApiOperation("FindSampleDomainById", tags: new[] { "SampleDomains" }, Description = "Find sample domain by id")]
-        [OpenApiParameter(name: "id", In = ParameterLocation.Path, Required = true, Type = typeof(Guid))]
-        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(SampleDomainDto))]
+        [FunctionName("FindSampleDomains")]
+        [OpenApiOperation("FindSampleDomains", tags: new[] { "SampleDomains" }, Description = "Find sample domains")]
+        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(List<SampleDomainDto>))]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(object))]
         public async Task<IActionResult> Run(
-            [HttpTrigger(AuthorizationLevel.Function, "get", Route = "sample-domains/{id}")] HttpRequest req,
-            Guid id,
+            [HttpTrigger(AuthorizationLevel.Function, "get", Route = "sample-domains")] HttpRequest req,
             CancellationToken cancellationToken)
         {
             try
             {
-                var result = await _appService.FindSampleDomainById(id);
+                var result = await _appService.FindSampleDomains(cancellationToken);
                 return new OkObjectResult(result);
             }
             catch (NotFoundException exception)