+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/UnMappedAzureFunction.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using AzureFunctions.TestApplication.Application.SampleDomains;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
 using AzureFunctions.TestApplication.Domain.Common.Interfaces;
@@ -41,7 +42,7 @@ namespace AzureFunctions.TestApplication.Api
             try
             {
                 var requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-                var request = JsonConvert.DeserializeObject<SampleDomainDto>(requestBody);
+                var request = JsonConvert.DeserializeObject<SampleDomainDto>(requestBody)!;
                 //IntentIgnore
                 return new NoContentResult();
             }